handle mermaid graph
