+++ src/preprocess/mod.rs
@@ -12,210 +12,14 @@ use crate::types::*;
 const BLOCK_DELIM: &str = "$$";
 const INLINE_BLOCK_DELIM: &str = "$";
 
+mod mermaid;
+pub use self::mermaid::*;
+
+mod format;
+pub use self::format::*;
+
 #[cfg(test)]
 mod tests;
-
-pub fn format_figure<'a>(
-    replacement: &Replacement<'a>,
-    refer: &str,
-    head_num: &str,
-    figures_counter: usize,
-    title: &str,
-    renderer: SupportedRenderer,
-) -> String {
-    use SupportedRenderer::*;
-    match renderer {
-        Html | Markdown => {
-            format!(
-                r#"<figure id="{refer}" class="figure">
-                    <object data="assets/{file}" type="image/svg+xml"/></object>
-                    <figcaption>Figure {head_num}{figures_counter} {title}</figcaption>
-                </figure>"#,
-                refer = refer,
-                head_num = head_num,
-                figures_counter = figures_counter,
-                title = title,
-                file = replacement.svg.display()
-            )
-        }
-        Latex | Tectonic => {
-            format!(r#"\[{}\]"#, replacement.intermediate())
-        }
-    }
-}
-
-pub fn format_equation_block<'a>(
-    replacement: &Replacement<'a>,
-    refer: &str,
-    head_num: &str,
-    equations_counter: usize,
-    renderer: SupportedRenderer,
-) -> String {
-    use SupportedRenderer::*;
-    match renderer {
-        Html | Markdown => {
-            format!(
-                r#"<div id="{refer}" class="equation">
-                    <div class="equation_inner">
-                        <object data="assets/{file}" type="image/svg+xml"></object>
-                    </div><span>({head_num}{equations_counter})</span>
-                </div>"#,
-                refer = refer,
-                head_num = head_num,
-                equations_counter = equations_counter,
-                file = replacement.svg.display()
-            )
-        }
-        Latex | Tectonic => {
-            format!(r#"\[{}\]"#, replacement.intermediate())
-        }
-    }
-}
-
-pub fn format_equation<'a>(replacement: &Replacement<'a>, renderer: SupportedRenderer) -> String {
-    use SupportedRenderer::*;
-    match renderer {
-        Html | Markdown => {
-            format!(
-                r#"<div class="equation"><div class="equation_inner"><object data="assets/{file}" type="image/svg+xml"></object></div></div>\n"#,
-                file = replacement.svg.display()
-            )
-        }
-        Latex | Tectonic => {
-            format!(r#"\[{}\]"#, replacement.intermediate())
-        }
-    }
-}
-
-pub fn format_inline_equation<'a>(
-    replacement: &Replacement<'a>,
-    renderer: SupportedRenderer,
-) -> String {
-    use SupportedRenderer::*;
-    match renderer {
-        Html | Markdown => {
-            format!(
-                r#"<object class="equation_inline" data="assets/{file}" type="image/svg+xml"></object>"#,
-                file = replacement.svg.display()
-            )
-        }
-        Latex | Tectonic => {
-            format!(r#"${}$"#, replacement.content.s)
-        }
-    }
-}
-
-fn create_svg_from_mermaid(
-    code: &str,
-    dest: impl AsRef<Path>,
-    chapterno: &str,
-    counter: usize,
-) -> Result<PathBuf> {
-    let mmdc = which::which("mmdc")?;
-    let dest = dest.as_ref();
-
-    let dest = dest.join(format!("mermaid_{}_{}.svg", chapterno, counter));
-
-    let mut child = std::process::Command::new(mmdc)
-        .arg("--outputFormat=svg")
-        .arg(format!("--output={}", dest.display()))
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .spawn()?;
-
-    // FIXME make this simpler
-    let code = code.to_owned();
-    let mut stdin = child.stdin.take().unwrap();
-    let j = std::thread::spawn(move || {
-        stdin.write(code.as_bytes())?;
-        Ok::<_, crate::errors::Error>(())
-    });
-    let mut stdout = child.stdout.unwrap();
-    let mut buf = String::with_capacity(8192);
-    stdout.read_to_string(&mut buf)?;
-
-    j.join().unwrap()?;
-    dbg!(buf);
-
-    Ok(dest)
-}
-
-/// Currently there is no way to display mermaid
-/// TODO FIXME
-pub fn gen_mermaid_charts(
-    source: &str,
-    chapterno: String,
-    dest: impl AsRef<Path>,
-    renderer: SupportedRenderer,
-) -> Result<String> {
-    match renderer {
-        // markdown and html can just fine deal with it
-        SupportedRenderer::Html => return Ok(source.to_owned()),
-        // SupportedRenderer::Markdown => return Ok(source.to_owned()),
-        _ => {
-            eprintln!("Stripping `mermaid` fencing of code block, not supported yet")
-        }
-    }
-
-    let dest = dest.as_ref();
-
-    use pulldown_cmark::*;
-    use pulldown_cmark_to_cmark::cmark;
-
-    let mut buf = String::with_capacity(source.len());
-
-    #[derive(Debug, Default)]
-    struct State {
-        is_mermaid_block: bool,
-        counter: usize,
-    }
-
-    let events = Parser::new_ext(&source, Options::all())
-        .into_offset_iter()
-        .scan(State::default(), |state, (mut event, _offset)| {
-            match event {
-                Event::Start(Tag::CodeBlock(ref mut kind)) => match kind {
-                    CodeBlockKind::Fenced(s) if s.as_ref() == "mermaid" => {
-                        *kind = CodeBlockKind::Fenced("text".into());
-                        state.counter += 1;
-                        state.is_mermaid_block = true;
-                        return None;
-                    }
-                    _ => {}
-                },
-                Event::End(Tag::CodeBlock(ref mut kind)) => match kind {
-                    CodeBlockKind::Fenced(s) if s.as_ref() == "mermaid" => {
-                        *kind = CodeBlockKind::Fenced("text".into());
-                        state.is_mermaid_block = false;
-                        return None;
-                    }
-                    _ => {}
-                },
-                Event::Code(ref code) => {
-                    if state.is_mermaid_block {
-                        let svg_path = dbg!(create_svg_from_mermaid(
-                            code.as_ref(),
-                            dest,
-                            chapterno.as_str(),
-                            state.counter
-                        )
-                        .expect("mermaid graph issue"));
-                        let inject = Tag::Image(
-                            LinkType::Inline,
-                            "url".into(),
-                            format!("Chapter {} Graphic {}", chapterno.as_str(), state.counter)
-                                .into(),
-                        );
-                    }
-                }
-                _ => {}
-            }
-            Some(event)
-        });
-
-    pulldown_cmark_to_cmark::cmark(events, &mut buf).map_err(Error::CommonMarkGlue)?;
-    Ok(buf)
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 struct SplitTagPosition<'a> {
